feat: add react intl universal file
Issue: https://github.com/alibaba/react-intl-universal/issues/115
