+++ package.json
@@ -51,11 +51,11 @@
     "babel-preset-es2015-rollup": "^3.0.0",
     "jest": "^19.0.2",
     "jest-cli": "^19.0.2",
+    "react": "^15.4.2",
     "rimraf": "^2.4.2",
     "rollup": "^0.41.4",
     "rollup-plugin-babel": "^2.3.9",
-    "rollup-plugin-commonjs": "^7.0.0",
-    "react": "^15.4.2"
+    "rollup-plugin-commonjs": "^7.0.0"
   },
   "babel": {
     "presets": [
